8.1.5 Add tests for all reducers and async actions.
